Move IntegralGenerator to codegeneration
